PowerDeathCommand Added, close #39, see full comment for usage /rpgitem [ItemName] power deathcommand [Chance] [Count] [Command] [Description] /rpgitem [ItemName] power deathcommand [Chance] [Command] // Count default to 1
This power, triggered with the chance 1/[Chance], kills the enemy immediately then execute the [Command] for [Count] times.
Use '`' to embrace the command if it contains spaces. `${x}` `${y}` `${z}` will be replaced by the coordinates where the enemy dies.

e.g.
/rpgitem somesword power deathcommand 100 3 `/summon Zombie ${x} ${y} ${z} {CustomName:BODY,CustomNameVisible:1}` `spawn three zombies`

Then the item `somesword` will have a 1% chance to kill the enemy then spawn three zombies.
The lore will be: "1/100 chance to kill the enemy then spawn three zombies"

The command could be long and the command block may helps.